Performance loss with Arduino-friendly representation

+++ vimana/v12_c_cons/base.h
@@ -13,8 +13,8 @@ Basic data types and functions.
 
 typedef unsigned char    VByte;
 typedef unsigned int     VUInt;
-//typedef unsigned long    VAddr;
-typedef unsigned int     VAddr;
+typedef unsigned long    VAddr;
+//typedef unsigned int     VAddr;
 typedef unsigned long    VULong;
 typedef          long    VIntNum;
 typedef          double  VDecNum;
